utils: Move OkHttpClientHelper to its own file
This concludes the effort so modularize the monolithic "Utils.kt" file.
